requirements.txt
# Database
psycopg2==2.9.9
SQLAlchemy==2.0.25

# Flask and extensions
Flask==2.3.3
Flask-Cors==4.0.0

# Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Google OAuth
google-auth==2.23.4

# Subscriptions
stripe==12.0.0
paypalrestsdk==1.13.3

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
